% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Epusillus}
\alias{Epusillus}
\title{Etmopterus pusillus data}
\format{A data frame with five variables:
\describe{
\item{\code{species}}{Full species name}
\item{\code{sex}}{Sex of the animal: Female (\code{F}) or Male (\code{M})}
\item{\code{age}}{Age in years}
\item{\code{length}}{Total length in centimetres}
\item{\code{maturity}}{Maturation status: \code{immature} or \code{mature}}
}}
\source{
Coelho, R. & Erzini, K. (2007). Population parameters of the smooth
lantern shark, \emph{Etmopterus pusillus}, in southern portugal (NE Atlantic). \emph{Fisheries Research}, 86, 42--57.
}
\usage{
Epusillus
}
\description{
Growth data for deepwater smooth lantern shark \emph{Etmopterus pusillus} from Coelho and Erzini (2007). Data are cross-sectional with one observation (row) per individual.
}
\keyword{datasets}
