% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Espinax}
\alias{Espinax}
\title{Etmopterus spinax data}
\format{A data frame with five variables:
\describe{
\item{\code{species}}{Full species name}
\item{\code{sex}}{Sex of the animal: Female (\code{F}) or Male (\code{M})}
\item{\code{age}}{Age in years}
\item{\code{length}}{Total length in centimetres}
\item{\code{maturity}}{Maturation status: \code{immature} or \code{mature}}
}}
\source{
Coelho, R. and Erzini, K. (2008). Life history of a wide-ranging deepwater lantern shark in the
north-east Atlantic, \emph{Etmopterus spinax} (chondrichthyes: Etmopteridae), with implications for
conservation. \emph{Journal of Fish Biology}, 73, 1419--1443.
}
\usage{
Espinax
}
\description{
Growth data for deepwater velvet belly lantern shark \emph{Etmopterus spinax} from Coelho and Erzini (2008). Data are cross-sectional with one observation (row) per individual.
}
\keyword{datasets}

