% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maturity_post_prob.R
\name{get_maturity_post_prob}
\alias{get_maturity_post_prob}
\title{Posterior probability of sex being female based on maturity alone}
\usage{
get_maturity_post_prob(mixprop, pF, pM, data)
}
\arguments{
\item{mixprop}{Numeric scalar of mixing proportion (overall sex ratio)}

\item{pF}{Numeric vector with predicted female maturity probability.}

\item{pM}{Numeric vector with predicted male maturity probability.}

\item{data}{A data.frame with binary column "maturity".}
}
\value{
Numeric vector of the posterior probability of being female.
}
\description{
\code{get_growth_post_prob} returns the probability of
the observation(s) arising from the female component given a set of parameters
and an assumed distribution (binomial). The component 
probability is given by Bayes' theorem. Used internally.
}
\examples{
get_maturity_post_prob(mixprop = 0.5, pF = 0.5, pM = 0.5, sigmaF = 1,
                          data = data.frame(maturity = 1))
}
