% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growth_mix.R
\name{growth_mix}
\alias{growth_mix}
\title{Fit finite mixture growth models.}
\usage{
growth_mix(start.fit, data, maxiter.em = 1000, abstol = 1e-08,
  plot.fit = TRUE, verbose = TRUE, estimate.mixprop = TRUE)
}
\arguments{
\item{start.fit}{A list with starting models for (optionally): "female_growth_fit", "male_growth_fit", "mixprop".}

\item{data}{A data.frame with columns: "age", "length" (for growth) and "obs.sex". "obs.sex" must have values "female", "male", "immature".}

\item{maxiter.em}{Integer for maximum number of EM iterations (1e3 default).}

\item{abstol}{Tolerance for EM observed data log likelihood convergence (1e-8 default).}

\item{plot.fit}{Logical, if TRUE fit plotted per iteration.}

\item{verbose}{Logical, if TRUE iteration and observed data log-likelihood printed.}

\item{estimate.mixprop}{Logical, if TRUE the mixing proportion is estimated, otherwise fixed at the starting value.}
}
\value{
List containing the components:
\item{ollike}{Observed data log-likelihood at each iteration.}
\item{complete_data}{Data frame of the data (re-ordered) with component probabilities (tau).}
\item{mixprop}{Estimate of the female mixing proportion.}
\item{par}{A list containing final complete data model estimates.}
\item{convergence}{Binary with a "0" denoting convergence.}
}
\description{
\code{growth_mix} fits sex-specific growth models where some of the animals are of unknown sex. Optimization is via the Expectation-Maximisation algorithm. Mean form for growth is flexible (tested on "nls", "gam" in package mgcv, "segmented" in package "segmented"). Assumes a normal distribution for growth currently.
}
\examples{
set.seed(1010)
sim.dat <- sim_vb_data(nfemale = 50, nmale = 50, mean_ageF = 4, mean_ageM = 4,
                      growth_parF = c(linf = 30, k = 0.5, t0 = -1, sigma = 1),
                      growth_parM = c(linf = 25, k = 0.5, t0 = -1, sigma = 1),
                      mat_parF = c(A50 = 5, MR = 2), mat_parM = c(A50 = 3, MR = 2),
                      distribution = "normal")

## set weights to one initially
sim.dat$weights <- 1
## Don't ask for each iteration plot
options(device.ask.default = FALSE)
## Additive growth model fit
start.fit<-list(
            female_growth_fit = mgcv::gam(length ~ s(age, k=5),
                                          data=subset(sim.dat, obs.sex=="female")),
            male_growth_fit = mgcv::gam(length ~ s(age, k=5),
                                        data=subset(sim.dat, obs.sex=="male")),
            mixprop = 0.5
            )

add.fit <- growth_mix(data = sim.dat, start.fit = start.fit)
options(device.ask.default = TRUE)
}

