% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity_mix.R
\name{maturity_mix}
\alias{maturity_mix}
\title{Fit finite mixture maturity model.}
\usage{
maturity_mix(start.list, data, binding = NULL, maxiter.em = 1000,
  reltol = 1e-08, plot.fit = FALSE, verbose = TRUE,
  estimate.mixprop = TRUE)
}
\arguments{
\item{start.list}{A list with a list called par containing starting values for: "mixprop", "betaF" (female parameters) and "betaM" (male parameters) (see Examples).}

\item{data}{A data.frame with columns: "maturity" (binary), "x" (single explanatory covariate currently), and "obs.sex". "obs.sex" must have values "female", "male", "unclassified".}

\item{binding}{A (kx2) parameter index matrix (not implemented yet)}

\item{maxiter.em}{Integer for maximum number of EM iterations (1e3 default).}

\item{reltol}{Relative tolerance for EM observed data log likelihood convergence (1e-8 default).}

\item{plot.fit}{Logical, if TRUE fit plotted per iteration (binary maturity jittered for visualisation). Red and blue circles are used for known females and males, respectively. Unclassified animals are plotted as triangle with the colour indicating the expected probability of being female or male (FALSE default).}

\item{verbose}{Logical, if TRUE iteration and observed data log-likelihood printed.}

\item{estimate.mixprop}{Logical, if TRUE the mixing proportion is estimated, otherwise fixed at the starting value.}
}
\value{
List containing the components:
\item{logLik.vec}{Observed data log-likelihood at each iteration.}
\item{logLik}{Observed data log-likelihood on the last EM iteration.}
\item{complete_data}{Data frame of the data (re-ordered) with component probabilities (tau).}
\item{coefficients}{Parameter estimates (on logit scale) by sex and associated standard errors (not currently implemented).}
\item{vcov}{Estimated variance covariance matrix of the parameters estimated on the real line. Can be used to obtain parameter standard errors on the natural scale (not currently implemented).}
\item{convergence}{Binary with a "0" denoting convergence of the EM algorithm.}
}
\description{
\code{maturity_mix} fits sex-specific maturity models where some of the animals are of unknown sex. Optimization is via the Expectation-Maximisation algorithm.
}
\examples{
set.seed(1011)
sim.dat<-sim_mat_data(nfemale = 30, nmale = 30, shapes = c(3, 3),
                     mat_parF = c(x50 = 0.5, xr = 0.1),
                     mat_parM = c(x50 = 0.4, xr = 0.1),
                     xcutoff = c(female = 0.4, male = 0.3))

start.list <- list(par = list(
                       mixprop = 0.5,
                       betaF = matrix(c(-5, 10)),
                       betaM = matrix(c(-5, 10))
                   ))

mat.fit <- maturity_mix(start.list = start.list, data = sim.dat)
}
