% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mat_data.R
\name{sim_mat_data}
\alias{sim_mat_data}
\title{Simulate sex-specific maturity data with missing classifications.}
\usage{
sim_mat_data(nfemale, nmale, shapes, mat_parF, mat_parM, xcutoff)
}
\arguments{
\item{nfemale}{Numeric scalar for number of female observations.}

\item{nmale}{Numeric scalar for number of male observations.}

\item{shapes}{Numeric vector for beta-distribution shape parameters - used to generate x values}

\item{mat_parF}{Named ("x50", "xr") numeric vector with female maturity parameters ("x50" is the x value at which 50\% are mature; "xr" is the x range between 25\% and 75\% mature)}

\item{mat_parM}{Named ("x50", "xr") numeric vector with male maturity parameters}

\item{xcutoff}{Named ("female", "male") numeric vector with x cutoffs below which immature animals are unclassified}
}
\value{
data.frame with columns "x", "true.sex", "obs.sex" (observed sex including unclassified), "maturity" (binary: 1 if mature; 0 if immature).
}
\description{
\code{sim_mat_data} simulates sex-specific maturity data according to
a logistic maturity model.
}
\examples{
sim.dat<-sim_mat_data(nfemale = 30, nmale = 30, shapes = c(3, 3),
                     mat_parF = c(x50 = 0.5, xr = 0.1),
                     mat_parM = c(x50 = 0.4, xr = 0.1),
                     xcutoff = c(female = 0.5, male = 0.4))
}
