% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vb_growth_mix.R
\name{vb_growth_mix}
\alias{vb_growth_mix}
\title{Fit finite mixture von Bertalanffy growth model.}
\usage{
vb_growth_mix(start.fit, data, binding, maxiter.em = 1000, abstol = 1e-09,
  plot.fit = TRUE, verbose = TRUE, optim.method = "BFGS",
  estimate.mixprop = TRUE, distribution)
}
\arguments{
\item{start.fit}{A list with a list called par containing starting values for: "mixprop", "growth.par" (see Examples).}

\item{data}{A data.frame with columns: "age", "length" and "obs.sex". "obs.sex" must have values "female", "male", "immature".}

\item{binding}{A (4x2) parameter index matrix with rows named (in order): "lnlinf", "lnk", "lnnt0", "lnsigma" and the left column for the female parameter index and right column for mal parameter index. Used to impose arbitrary equality constraints across the sexes (see Examples).}

\item{maxiter.em}{Integer for maximum number of EM iterations (1e3 default).}

\item{abstol}{Tolerance for EM observed data log likelihood convergence (1e-9 default).}

\item{plot.fit}{Logical, if TRUE fit plotted per iteration. Red and blue circles are used for known females and males, respectively. Immature / unsexed animals are plotted as triangle with the colour indicating the expected probability of being female or male.}

\item{verbose}{Logical, if TRUE iteration and observed data log-likelihood printed.}

\item{optim.method}{Character, complete data optimisation method to use in \code{optim}.}

\item{estimate.mixprop}{Logical, if TRUE the mixing proportion is estimated, otherwise fixed at the starting value.}

\item{distribution}{Character with options: "normal" or "lognormal".}
}
\value{
List containing the components:
\item{logLik.vec}{Observed data log-likelihood at each iteration.}
\item{logLik}{Observed data log-likelihood on the last EM iteration.}
\item{complete_data}{Data frame of the data (re-ordered) with component probabilities (tau).}
\item{coefficients}{Parameter estimates (on the real line) and associated standard errors on the real line.}
\item{vcov}{Estimated variance covariance matrix of the parameters estimated on the real line. Can be used to obtain parameter standard errors on the natural scale.}
\item{convergence}{Binary with a "0" denoting convergence of the EM algorithm.}
}
\description{
\code{vb_growth_mix} fits sex-specific growth models where some of the animals are of unknown sex. Optimization is via the Expectation-Maximisation algorithm. Equality constraints across sexes can be implemented for any combination of parameters using the \code{binding} argument. Assumes a normal distribution currently.
}
\examples{
set.seed(1010)
sim.dat <- sim_vb_data(nfemale = 50, nmale = 50, mean_ageF = 4, mean_ageM = 4,
                      growth_parF = c(linf = 30, k = 0.5, t0 = -1, sigma = 0.1),
                      growth_parM = c(linf = 25, k = 0.5, t0 = -1, sigma = 0.1),
                      mat_parF = c(A50 = 5, MR = 2), mat_parM = c(A50 = 3, MR = 2),
                      distribution = "lognormal")

## Model fit with contrained Brody's growth coefficient
## Set up the constraint
binding <- matrix(c(1:2, rep(3, 2), 4:7), ncol = 2, byrow = TRUE)
rownames(binding) <- c("lnlinf", "lnk", "lnnt0", "lnsigma")
colnames(binding) <- c("female", "male")
## starting values
start.par <- c(rep(log(25), 2), rep(log(0.2), 1), rep(log(3), 2), rep(log(1), 2))
start.list <- list(par = list(mixprop = 0.5, growth.par = start.par))
## Don't ask for each iteration plot
options(device.ask.default = FALSE)
vb.bind.fit <- vb_growth_mix(data = sim.dat, start.fit = start.list,
                             binding = binding, distribution = "lognormal")
options(device.ask.default = TRUE)
}

