% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vb_lengths.R
\name{vb_lengths}
\alias{vb_lengths}
\title{von Bertalanffy growth function.}
\usage{
vb_lengths(theta, age)
}
\arguments{
\item{theta}{A numeric vector with named values "linf", "k", "t0".}

\item{age}{A numeric vector of ages.}
}
\value{
Predicted length-at-age.
}
\description{
\code{vb_lengths} returns the predicted length-at-age for given named set of parameters for the von Bertalanffy growth function:
\deqn{L=L_{\infty}(1-e^{-k(A-t_0)})}
}
\examples{
vb_lengths(theta=c("linf"=30,"k"=0.2,"t0"=-1), age=0:10)
}

